// src/v2/components/native-select/types.ts
import { ChangeEvent, HTMLAttributes, ReactNode, SelectHTMLAttributes } from 'react';
import type { TInputSize, TInputVariant } from '../input/input-wrapper/types';
import type { TPaletteColor } from '../../../theme/types';
import type { TFormControlChangeHandler } from '../form-control/types';

export type TNativeSelectOption = {
  value: string;
  label: ReactNode;
  disabled?: boolean;
};

export type TNativeSelectChangeHandler = TFormControlChangeHandler<string>;

export type TNativeSelectProps = Omit<
  HTMLAttributes<HTMLDivElement>,
  'color' | 'onChange' | 'defaultValue'
> & {
  options: Array<string | TNativeSelectOption>;
  value?: string;
  defaultValue?: string;
  onChange?: TNativeSelectChangeHandler;
  placeholder?: string;
  name?: SelectHTMLAttributes<HTMLSelectElement>['name'];
  autoComplete?: SelectHTMLAttributes<HTMLSelectElement>['autoComplete'];
  label?: ReactNode;
  description?: ReactNode;
  helperText?: ReactNode;
  required?: boolean;
  error?: boolean;
  fullWidth?: boolean;
  disabled?: boolean;
  size?: TInputSize;
  variant?: TInputVariant;
  color?: TPaletteColor;
  actionBar?: ReactNode;
};